# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Battleship Backend API
  version: 1.0.0
servers:
  - url: http://localhost:5175
tags:
  - name: Auth
  - name: Users
  - name: Stats
  - name: Leaderboard
paths:
  /health:
    get:
      summary: Health check
      tags: [Users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /api/auth/register:
    post:
      summary: Register a user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing fields
        "409":
          description: User exists
  /api/auth/login:
    post:
      summary: Log in
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing fields
        "401":
          description: Invalid credentials
  /api/users/me:
    get:
      summary: Get current user
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
  /api/users/{userId}/stats:
    get:
      summary: Get user stats
      tags: [Stats]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatsResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Stats not found
    put:
      summary: Create user stats
      tags: [Stats]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatsRequest"
      responses:
        "201":
          description: Created stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatsResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Stats already exists
    patch:
      summary: Update user stats
      tags: [Stats]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatsRequest"
      responses:
        "200":
          description: Updated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatsResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Stats not found
  /api/leaderboard:
    get:
      summary: Get leaderboard
      tags: [Leaderboard]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  leaderboard:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeaderboardEntry"
        "401":
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    UserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
    UserStatsRequest:
      type: object
      properties:
        gamesPlayed:
          type: integer
          minimum: 0
        wins:
          type: integer
          minimum: 0
        losses:
          type: integer
          minimum: 0
        hits:
          type: integer
          minimum: 0
        misses:
          type: integer
          minimum: 0
        shotsFired:
          type: integer
          minimum: 0
    UserStatsResponse:
      type: object
      properties:
        stats:
          $ref: "#/components/schemas/UserStats"
    UserStats:
      type: object
      properties:
        userId:
          type: string
        gamesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        hits:
          type: integer
        misses:
          type: integer
        shotsFired:
          type: integer
    LeaderboardEntry:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        gamesPlayed:
          type: integer
        hits:
          type: integer
        misses:
          type: integer
        shotsFired:
          type: integer